+++ Mods ChoGGi/Change Other To Random Gender/metadata.lua
@@ -1,11 +1,10 @@
 return PlaceObj("ModDef", {
 	"title", "Change Other To Random Gender",
+	"id", "ChoGGi_ChangeOtherToRandomGender",
 	"version", 1,
 	"version_major", 0,
 	"version_minor", 1,
-
 	"image", "Preview.png",
-	"id", "ChoGGi_ChangeOtherToRandomGender",
 	"steam_id", "1541604722",
 	"pops_any_uuid", "bccb41f2-3172-48ba-8466-ca113ce458aa",
 	"author", "ChoGGi",
@@ -13,8 +12,13 @@ return PlaceObj("ModDef", {
 	"code", {
 		"Code/Script.lua",
 	},
-	"description", [[Any colonists born/newly landed on mars with the Other gender will be randomly changed into Male/Female.
+	"description", [[
+Community:CalcBirth() only uses Male/Female genders to generate births, Other gender isn't included in the fertility count.
+Use this mod to maximize birth rates on higher difficulty maps.
+
+This will change any colonists born/newly landed on mars with the Other gender to be randomly changed into Male/Female.
 
 
-Requested by Olaf.]],
+Requested by Olaf.
+]],
 })