+++ Mods ChoGGi/Exotic Meteors/metadata.lua
@@ -1,0 +1,27 @@
+return PlaceObj("ModDef", {
+	"title", "Exotic Meteors",
+	"id", "ChoGGi_ExoticMeteors",
+	"steam_id", "3274108477",
+	"pops_any_uuid", "c829c12b-39ba-472c-bd54-e53b08fa5c27",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Meteors can drop rare metals/exotic minerals.
+Only works on B&B (for now, but no idea why it doesn't work without B&B).
+
+
+See mod options to adjust drop rates.
+
+
+Requested by Anete.
+]],
+})